Тест нижней кнопки Заказать вынесен в отдельный класс
